requirements.txt
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
langchain_groq
crewai
crewai_tools
python-dotenv
langchain_chroma
sentence-transformer